Split publisher main
